+++ src/bin/platforms/linux.ts
@@ -216,12 +216,12 @@ export class LinuxInstaller {
   private checkForRoot() {
     if (process.getuid() !== 0) {
       this.hbService.logger('ERROR: This command must be executed using sudo on Linux', 'fail');
-      this.hbService.logger(`sudo hb-service ${this.hbService.action} --user ${this.hbService.asUser || 'your-user'}`, 'fail');
+      this.hbService.logger(`EXAMPLE: sudo hb-service ${this.hbService.action} --user ${this.hbService.asUser || 'your-user'}`, 'fail');
       process.exit(1);
     }
     if (this.hbService.action === 'install' && !this.hbService.asUser) {
       this.hbService.logger('ERROR: User parameter missing. Pass in the user you want to run Homebridge as using the --user flag eg.', 'fail');
-      this.hbService.logger(`sudo hb-service ${this.hbService.action} --user your-user`, 'fail');
+      this.hbService.logger(`EXAMPLE: sudo hb-service ${this.hbService.action} --user your-user`, 'fail');
       process.exit(1);
     }
   }
@@ -350,6 +350,7 @@ export class LinuxInstaller {
       `Type=simple`,
       `User=${this.hbService.asUser}`,
       `PermissionsStartOnly=true`,
+      `WorkingDirectory=${this.hbService.storagePath}`,
       `EnvironmentFile=/etc/default/${this.systemdServiceName}`,
       `ExecStartPre=-run-parts ${this.runPartsPath}`,
       `ExecStartPre=-${this.hbService.selfPath} before-start $HOMEBRIDGE_OPTS`,